Projects/Data_Pipeline_with_GCP/function/requirements.txt
#Update your requirements.txt with the following dependencies
requests_toolbelt==0.9.1
google-auth==1.27.0